API request to pull down picture of the day
Added API ket in the env file so it is not public on github, then referanced this Key in the request and added it in the runtimeConfig to access it.  Created apod.js which uses cachedEventHandler to be able to add a maxAge of 12 hours to the request and a key so it is unique cached key, the file gets the API key and adds it in the fetch request and returns just the data we want from the request.  Created pictureOfTheDay component which uses useFetch to get the request data and then I used NuxtImg to render the image, if the hdImage is there use that otherwise use the smaller image, nuxtImg will auto create the appropriate webp image for us at the sizes I provided and lazy load in the image.  I have added the title and description below the image.